+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - Fix conversation cleanup when MessagingClientConversationJoinFailure occurs to prevent orphaned conversations
 
+### Changed
+
+- Clean up `ChatSDK.sendTypingEvent()` unused code path
+
 ## [1.11.2] - 2025-06-24
 
 ### Changed