+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.11.3]
+
 ### Fixed
 
 - Fix conversation cleanup when MessagingClientConversationJoinFailure occurs to prevent orphaned conversations